+++ stag_lib/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the library are documented here.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Reduced the memory usage of the SBM methods
+
+### Added
+- Added signless Laplacian methods to the graph object
+
 ## [1.1.0] - 2023-02-23
 ### Added
 - [Issue #70](https://github.com/staglibrary/stag/issues/70): Add general stochastic block model method